# openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MeisterProPR.Api",
    "version": "1.0"
  },
  "paths": {
    "/admin/crawl-configurations": {
      "get": {
        "tags": [
          "AdminCrawlConfigs"
        ],
        "summary": "Lists all accessible crawl configurations.\r\nAdmins receive all configs; non-Admin users receive only configs for their assigned clients.",
        "responses": {
          "200": {
            "description": "List of crawl configurations.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrawlConfigResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrawlConfigResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrawlConfigResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials provided.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminCrawlConfigs"
        ],
        "summary": "Creates a new crawl configuration.\r\nAdmins may create for any client; non-Admin users may only create for their assigned clients.",
        "requestBody": {
          "description": "Crawl configuration details.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminCrawlConfigRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminCrawlConfigRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminCrawlConfigRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Configuration created.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlConfigResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlConfigResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlConfigResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials provided.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Non-Admin caller does not own the specified client.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Client not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Config for this org/project already exists for the client.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/admin/crawl-configurations/{configId}": {
      "patch": {
        "tags": [
          "AdminCrawlConfigs"
        ],
        "summary": "Applies partial updates to a crawl configuration.\r\nAdmins may update any config; non-Admin users may only update configs for their clients.",
        "parameters": [
          {
            "name": "configId",
            "in": "path",
            "description": "Configuration identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Fields to update; omit a field to leave it unchanged.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAdminCrawlConfigRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAdminCrawlConfigRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAdminCrawlConfigRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuration updated.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlConfigResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlConfigResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlConfigResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials provided.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Non-Admin caller does not own this config's client.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Configuration not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AdminCrawlConfigs"
        ],
        "summary": "Deletes a crawl configuration.\r\nAdmins may delete any config; non-Admin users may only delete configs for their clients.",
        "parameters": [
          {
            "name": "configId",
            "in": "path",
            "description": "Configuration identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Configuration deleted."
          },
          "401": {
            "description": "No valid credentials provided.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Non-Admin caller does not own this config's client.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Configuration not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CrawlConfigResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "organizationUrl": {
            "type": "string",
            "nullable": true
          },
          "projectId": {
            "type": "string",
            "nullable": true
          },
          "crawlIntervalSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "repoFilters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CrawlRepoFilterResponse"
            },
            "description": "Repo and branch filters. Empty array means all repos are crawled."
          }
        },
        "additionalProperties": false,
        "description": "Crawl configuration response."
      },
      "CreateAdminCrawlConfigRequest": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "organizationUrl": {
            "type": "string",
            "nullable": true
          },
          "projectId": {
            "type": "string",
            "nullable": true
          },
          "crawlIntervalSeconds": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Request body for creating an admin-managed crawl configuration."
      },
...
      "PatchAdminCrawlConfigRequest": {
        "type": "object",
        "properties": {
          "crawlIntervalSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "nullable": true
          },
          "repoFilters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CrawlRepoFilterRequest"
            },
            "nullable": true,
            "description": "Full replacement of the repo filter list. Omit to leave unchanged; send [] to clear all filters."
          }
        },
        "additionalProperties": false,
        "description": "Request body for patching an admin-managed crawl configuration.\r\nOmit a field to leave it unchanged."
      },
      "CrawlRepoFilterResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "repositoryName": {
            "type": "string",
            "nullable": true
          },
          "targetBranchPatterns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Glob patterns for target branch names. Empty array means all branches accepted."
          }
        },
        "additionalProperties": false,
        "description": "Repo filter entry in a crawl configuration response."
      },
      "CrawlRepoFilterRequest": {
        "type": "object",
        "properties": {
          "repositoryName": {
            "type": "string",
            "nullable": true
          },
          "targetBranchPatterns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Glob patterns for target branch names. Empty array means all branches accepted."
          }
        },
        "additionalProperties": false,
        "description": "Repo filter entry in a PATCH crawl configuration request."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}